added route, fixed text, updates inline, svg
